pyproject.toml
[project]
name = "sombraaisquad-diia-aicontest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.27",
]
